Fix stat metrics accounting. Refactor setting metrics to be more general.
